requirements.txt
hydrus-api
pillow
pyocr
python-dotenv
requests
